*** empty log message ***
git-svn-id: svn://svn.code.sf.net/p/iup/iup/trunk@2203 2f08bc20-984a-4df8-b195-9b5d5d477537

+++ iup/srcglcontrols/iup_glcontrols.h
@@ -39,11 +39,11 @@ void iupGLFontFinish(void);
 void iupGLFontRelease(Ihandle* gl_parent);
 void iupGLFontGetDim(Ihandle* ih, int *maxwidth, int *height, int *ascent, int *descent);
 
-void iupGLDrawText(Ihandle* ih, double x, double y, const char* str, const char* color, int active);
-void iupGLDrawImage(Ihandle* ih, double x, double y, const char* name, int active);
-void iupGLDrawLine(Ihandle* ih, double x1, double y1, double x2, double y2, float width, const char* color, int active);
-void iupGLDrawRect(Ihandle* ih, double xmin, double xmax, double ymin, double ymax, float width, const char* color, int active);
-void iupGLDrawBox(Ihandle* ih, double xmin, double xmax, double ymin, double ymax, const char* color);
+void iupGLDrawText(Ihandle* ih, int x, int y, const char* str, const char* color, int active);
+void iupGLDrawImage(Ihandle* ih, int x, int y, const char* name, int active);
+void iupGLDrawLine(Ihandle* ih, int x1, int y1, int x2, int y2, float width, const char* color, int active);
+void iupGLDrawRect(Ihandle* ih, int xmin, int xmax, int ymin, int ymax, float width, const char* color, int active);
+void iupGLDrawBox(Ihandle* ih, int xmin, int xmax, int ymin, int ymax, const char* color);
 
 void iupGLIconRegisterAttrib(Iclass* ic);
 void iupGLIconDraw(Ihandle* ih, int icon_width, int icon_height, const char *image, const char* title, const char* fgcolor, int active);
